Changing webapp structure by using yo/grunt/bower
Grunt will be a big support for starting and mainting the nodejs part,
while bower supports the building and deployment of the webapp to the
server.

Last but not least; all these changes should help in building the client
on angularJs. Therefore jade and so on will not be used and probably
removed with the next check-in.

Next to these bigger changes; a few whitespace problems were solved, and
the logger can be finally resolved by changing it to package winston.