Connected mongodb with view via nodejs api
NodeJS server now serves articles by providing an api responding to get
and post to /api/articles;
Returned articles will be filtered, that only title and content will be
returned.

Angular logic was extended to respond to this new functionality. Then
most of the dummy code could be removed, and this code can be used in
production. The data limit cap exists on the mongodb, and this drops
data if used more than 512MB.

Was has to be done?!
TESTS!!!!!!
styling....