Fixed first styling errors
Page content is now aligned; there is a top navigation bar and a sidebar
Both can be used for choosing an article. However, currently both have
the same content and the same articles; because it is a duplicated menu.

For this a navigation template has to be created for changing side/top
orientation, depending on article and sub-article.

Still missing: tests for the content. The changes could not be tracked
and new errors where maybe created.