Added functionality to delete articles
Client code has the ability to send delete events to the server, this
responds with 204 if the article exists, and an answer will always be
the found and deleted article for confirmation.

Furthermore, the server logs now each request.

Another change is the simplified Article schema; it contains now updated
and created_at timestamp; depending if its new or newly changed.