Adding https support to webapp project
For this a private key and certificate were generated and added
to the expressjs configuration. To listening servers have to run in
parallel. The first insecure server redirect all requests to the secure
server, and the secure server ignores the function if the request is
secure, otherwise the request gets rerouted again to the secure server.

For two running servers two different ports have to be used, therefore
the config has to contain a normal port and and a secure port.