Added first unit testing framework
Karma and jasmine helps in writing and executing simple unit tests with
a few keywords.

Currently chrome is started solely, but more to come if the computer
supports that much ram and that many browsers at the same time.

For the first tests the ContentCtrl should be testet and checked for
vulnerabilites. Especially the error checks will be very important if
the server cannot be reached or the article requests and posting will
take longer than expected (e.g. deletion of articles and the way how to
handle erros during this process).