Added a first authentication module to the webapp
First of all a lot of testcases were written for the authorisation
service and the cryptopgraphic controller. The controller should use
the encapsulated functionality how to encrypt / decrypt / hash of
inputs. The controller will mostly combine user input and the comm to
the server. However, it is still unclear, how to cascade not to be seen
content for unauthorized persons. Currently a route change would track
down, if the current user is authorized or not. However, this could be
localy changed to always true, however, the communication with the
server would deny the requests of an unauthorized client.