Adding authentication directive with tests
This commit adds the authentication directive which currently looks
very simple with only a login-element, which uses mostly the cryptctrl
and simple click handler for the click on login/logout, however, only
the toggle between login-logout is currently tested in one direction.

Goal will be the login and logout of users using the cryptctrl. This
was also the first test for directives, which had to be handled using
the karma-ng-html2js-preprocessor and had to be handled in the karma
test itselfs, which was more complicated than testing a controller or
service.