Introducing the routing part of the webapp
Now the real articles module was linked and is shown via the view
using the previously defined article service and the responding server.

Next to this, the message box, its directive and the linkings were moved
as well to the new 'modules' folder. Maybe the modules will be splitted
in more than one file if they get longer and more complicated.

Goal of this commit is a robust and better switchable view if other
parts should be shown than always the same articles view, e.g. after a
login some contacts, dates, messages or whatever, but not articles...