UI-Routes with templates and views is fixed
The actual resulting index.html however still looks like chaos.
However, the views are now organized like a tree and changed while
clicking on links. This change will help in organizing the views on the
client side.

A last commit is however needed to establish a real user experience
before pushing it to heroku.